feat: "no-public-class-fields" rule to "no-construct-in-public-property-of-construct"
