+++ src/__tests__/no-construct-in-public-property-of-construct.test.ts
@@ -1,6 +1,6 @@
 import { RuleTester } from "@typescript-eslint/rule-tester";
 
-import { noPublicClassFields } from "../rules/no-public-class-fields";
+import { noConstructInPublicPropertyOfConstruct } from "../rules/no-construct-in-public-property-of-construct";
 
 const ruleTester = new RuleTester({
   languageOptions: {
@@ -12,20 +12,23 @@ const ruleTester = new RuleTester({
   },
 });
 
-ruleTester.run("no-public-class-fields", noPublicClassFields, {
-  valid: [
-    // WHEN: field type is primitive
-    {
-      code: `
+ruleTester.run(
+  "no-construct-in-public-property-of-construct",
+  noConstructInPublicPropertyOfConstruct,
+  {
+    valid: [
+      // WHEN: field type is primitive
+      {
+        code: `
       class Construct {}
       class TestClass extends Construct {
         public test: string;
       }
       `,
-    },
-    // WHEN: field type is interface
-    {
-      code: `
+      },
+      // WHEN: field type is interface
+      {
+        code: `
       class Construct {}
       interface ITest {
         value: string;
@@ -34,10 +37,10 @@ ruleTester.run("no-public-class-fields", noPublicClassFields, {
         public test: ITest;
       }
       `,
-    },
-    // WHEN: field type is type alias
-    {
-      code: `
+      },
+      // WHEN: field type is type alias
+      {
+        code: `
       class Construct {}
       type TestType = {
         value: string;
@@ -46,90 +49,91 @@ ruleTester.run("no-public-class-fields", noPublicClassFields, {
         public test: TestType;
       }
       `,
-    },
-    // WHEN: field is private
-    {
-      code: `
+      },
+      // WHEN: field is private
+      {
+        code: `
       class Construct {}
       class DependencyClass {}
       class TestClass extends Construct {
         private test: DependencyClass;
       }
       `,
-    },
-    // WHEN: field has no type annotation
-    {
-      code: `
+      },
+      // WHEN: field has no type annotation
+      {
+        code: `
       class Construct {}
       class TestClass extends Construct {
         public test;
       }
       `,
-    },
-    // WHEN: field is protected
-    {
-      code: `
+      },
+      // WHEN: field is protected
+      {
+        code: `
       class Construct {}
       class DependencyClass {}
       class TestClass extends Construct {
         protected test: DependencyClass;
       }
       `,
-    },
-    // WHEN: super class is not a construct
-    {
-      code: `
+      },
+      // WHEN: super class is not a construct
+      {
+        code: `
       class DependencyClass {}
       class TestClass {
         public test: DependencyClass;
       }
       `,
-    },
-  ],
-  invalid: [
-    // WHEN: public field uses class type
-    {
-      code: `
+      },
+    ],
+    invalid: [
+      // WHEN: public field uses class type
+      {
+        code: `
       class Construct {}
       class DependencyClass {}
       class TestClass extends Construct {
         public test: DependencyClass;
       }
       `,
-      errors: [{ messageId: "noPublicClassFields" }],
-    },
-    // WHEN: implicitly public field uses class type
-    {
-      code: `
+        errors: [{ messageId: "noConstructInPublicPropertyOfConstruct" }],
+      },
+      // WHEN: implicitly public field uses class type
+      {
+        code: `
       class Construct {}
       class DependencyClass {}
       class TestClass extends Construct {
         test: DependencyClass;
       }
       `,
-      errors: [{ messageId: "noPublicClassFields" }],
-    },
-    // WHEN: readonly public field uses class type
-    {
-      code: `
+        errors: [{ messageId: "noConstructInPublicPropertyOfConstruct" }],
+      },
+      // WHEN: readonly public field uses class type
+      {
+        code: `
       class Construct {}
       class DependencyClass {}
       class TestClass extends Construct {
         public readonly test: DependencyClass;
       }
       `,
-      errors: [{ messageId: "noPublicClassFields" }],
-    },
-    // WHEN: constructor parameter property uses class type
-    {
-      code: `
+        errors: [{ messageId: "noConstructInPublicPropertyOfConstruct" }],
+      },
+      // WHEN: constructor parameter property uses class type
+      {
+        code: `
       class Construct {}
       class DependencyClass {}
       class TestClass extends Construct {
         constructor(public test: DependencyClass) {}
       }
       `,
-      errors: [{ messageId: "noPublicClassFields" }],
-    },
-  ],
-});
+        errors: [{ messageId: "noConstructInPublicPropertyOfConstruct" }],
+      },
+    ],
+  }
+);
