+++ src/index.ts
@@ -4,12 +4,12 @@ import { name, version } from "../package.json";
 
 import { constructConstructorProperty } from "./rules/construct-constructor-property";
 import { noConstructInInterface } from "./rules/no-construct-in-interface";
+import { noConstructInPublicPropertyOfConstruct } from "./rules/no-construct-in-public-property-of-construct";
 import { noConstructStackSuffix } from "./rules/no-construct-stack-suffix";
 import { noImportPrivate } from "./rules/no-import-private";
 import { noMutablePropertyOfPropsInterface } from "./rules/no-mutable-property-of-props-interface";
 import { noMutablePublicPropertyOfConstruct } from "./rules/no-mutable-public-property-of-construct";
 import { noParentNameConstructIdMatch } from "./rules/no-parent-name-construct-id-match";
-import { noPublicClassFields } from "./rules/no-public-class-fields";
 import { noVariableConstructId } from "./rules/no-variable-construct-id";
 import { pascalCaseConstructId } from "./rules/pascal-case-construct-id";
 import { propsNameConvention } from "./rules/props-name-convention";
@@ -18,20 +18,21 @@ import { requirePassingThis } from "./rules/require-passing-this";
 import { requirePropsDefaultDoc } from "./rules/require-props-default-doc";
 
 const rules = {
-  "no-construct-in-interface": noConstructInInterface,
-  "no-construct-stack-suffix": noConstructStackSuffix,
-  "no-parent-name-construct-id-match": noParentNameConstructIdMatch,
-  "no-public-class-fields": noPublicClassFields,
-  "pascal-case-construct-id": pascalCaseConstructId,
-  "require-passing-this": requirePassingThis,
-  "no-variable-construct-id": noVariableConstructId,
-  "no-mutable-public-property-of-construct": noMutablePublicPropertyOfConstruct,
-  "no-mutable-property-of-props-interface": noMutablePropertyOfPropsInterface,
   "construct-constructor-property": constructConstructorProperty,
-  "require-jsdoc": requireJSDoc,
-  "require-props-default-doc": requirePropsDefaultDoc,
-  "props-name-convention": propsNameConvention,
+  "no-construct-in-interface": noConstructInInterface,
+  "no-construct-in-public-property-of-construct":
+    noConstructInPublicPropertyOfConstruct,
+  "no-construct-stack-suffix": noConstructStackSuffix,
   "no-import-private": noImportPrivate,
+  "no-mutable-property-of-props-interface": noMutablePropertyOfPropsInterface,
+  "no-mutable-public-property-of-construct": noMutablePublicPropertyOfConstruct,
+  "no-parent-name-construct-id-match": noParentNameConstructIdMatch,
+  "no-variable-construct-id": noVariableConstructId,
+  "pascal-case-construct-id": pascalCaseConstructId,
+  "props-name-convention": propsNameConvention,
+  "require-jsdoc": requireJSDoc,
+  "require-passing-this": requirePassingThis,
+  "require-props-default-doc": requirePropsDefaultDoc,
 };
 
 const cdkPlugin = {
@@ -55,39 +56,39 @@ const createFlatConfig = (rules: Record<string, unknown>) => {
 };
 
 const recommended = createFlatConfig({
+  "cdk/construct-constructor-property": "error",
   "cdk/no-construct-in-interface": "error",
+  "cdk/no-construct-in-public-property-of-construct": "error",
   "cdk/no-construct-stack-suffix": "error",
+  "cdk/no-mutable-property-of-props-interface": "warn",
+  "cdk/no-mutable-public-property-of-construct": "warn",
   "cdk/no-parent-name-construct-id-match": [
     "error",
     { disallowContainingParentName: false },
   ],
-  "cdk/no-public-class-fields": "error",
+  "cdk/no-variable-construct-id": "error",
   "cdk/pascal-case-construct-id": "error",
   "cdk/require-passing-this": ["error", { allowNonThisAndDisallowScope: true }],
-  "cdk/no-variable-construct-id": "error",
-  "cdk/no-mutable-public-property-of-construct": "warn",
-  "cdk/no-mutable-property-of-props-interface": "warn",
-  "cdk/construct-constructor-property": "error",
 });
 
 const strict = createFlatConfig({
+  "cdk/construct-constructor-property": "error",
   "cdk/no-construct-in-interface": "error",
+  "cdk/no-construct-in-public-property-of-construct": "error",
   "cdk/no-construct-stack-suffix": "error",
+  "cdk/no-import-private": "error",
+  "cdk/no-mutable-property-of-props-interface": "error",
+  "cdk/no-mutable-public-property-of-construct": "error",
   "cdk/no-parent-name-construct-id-match": [
     "error",
     { disallowContainingParentName: true },
   ],
-  "cdk/no-public-class-fields": "error",
-  "cdk/pascal-case-construct-id": "error",
-  "cdk/require-passing-this": "error",
   "cdk/no-variable-construct-id": "error",
-  "cdk/no-mutable-public-property-of-construct": "error",
-  "cdk/no-mutable-property-of-props-interface": "error",
-  "cdk/construct-constructor-property": "error",
-  "cdk/require-jsdoc": "error",
-  "cdk/require-props-default-doc": "error",
+  "cdk/pascal-case-construct-id": "error",
   "cdk/props-name-convention": "error",
-  "cdk/no-import-private": "error",
+  "cdk/require-jsdoc": "error",
+  "cdk/require-passing-this": "error",
+  "cdk/require-props-default-doc": "error",
 });
 
 const configs = {
